Separating definitions from main, TUI running in parallel with endpoints addr passed in
